rename & README.md revision Undone: python generator
The code in model.py uses a Python generator, if needed, to generate data for training rather than storing the training data in memory. The model.py code is clearly organized and comments are included where needed.

As required by the rubric a Python generator must be used to prevent all data being loaded into memory. Have a look overhere for an explanation of generators.